### 2023-09-11 33d8751
Update paper.bib
Add DOI and fix references.
