+++ paper/paper.bib
@@ -1,13 +1,16 @@
 @article{althoff2023conform,
   title={Conform With the Wind},
   author={Althoff, Simon},
-  year={2023}
+  school={Lund University},
+  year={2023},
+  type={Master's Thesis}
 }
 
 @article{athey2019generalized,
   title={Generalized random forests},
   author={Athey, Susan and Tibshirani, Julie and Wager, Stefan},
-  year={2019}
+  year={2019},
+  doi={10.1214/18-aos1709}
 }
 
 @article{behnel2010cython,
@@ -18,7 +21,8 @@
   number={2},
   pages={31--39},
   year={2010},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/mcse.2010.118}
 }
 
 @article{biau2016random,
@@ -28,7 +32,8 @@
   volume={25},
   pages={197--227},
   year={2016},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11749-016-0481-7}
 }
 
 @article{breiman2001random,
@@ -38,7 +43,8 @@
   volume={45},
   pages={5--32},
   year={2001},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1023/A:1010933404324}
 }
 
 @article{cordoba2021spatially,
@@ -48,7 +54,8 @@
   volume={289},
   pages={112509},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jenvman.2021.112509}
 }
 
 @article{dean2022quantile,
@@ -59,7 +66,8 @@
   number={4},
   pages={682--709},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10729-022-09609-0}
 }
 
 @article{fang2018quantile,
@@ -70,18 +78,19 @@
   number={10},
   pages={e0205155},
   year={2018},
-  publisher={Public Library of Science San Francisco, CA USA}
+  doi={10.1371/journal.pone.0205155}
 }
 
 @article{francke2008estimation,
   title={Estimation of suspended sediment concentration and yield using linear models, random forests and quantile regression forests},
   author={Francke, T and L{\'o}pez-Taraz{\'o}n, JA and Schr{\"o}der, B},
-  journal={Hydrological Processes: An International Journal},
+  journal={Hydrological Processes},
   volume={22},
   number={25},
   pages={4892--4904},
   year={2008},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/hyp.7110}
 }
 
 @article{gyamerah2020long,
@@ -91,7 +100,8 @@
   volume={2020},
   pages={1--11},
   year={2020},
-  publisher={Hindawi Limited}
+  publisher={Hindawi},
+  doi={10.1155/2020/1972962}
 }
 
 @article{hengl2018random,
@@ -101,33 +111,38 @@
   volume={6},
   pages={e5518},
   year={2018},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ},
+  doi={10.7717/peerj.5518}
 }
 
 @book{koenker2005quantile,
   title={Quantile regression},
   author={Koenker, Roger},
-  volume={38},
   year={2005},
-  publisher={Cambridge university press}
+  series={Econometric Society Monographs},
+  publisher={Cambridge University Press},
+  doi={10.1017/CBO9780511754098}
 }
 
 @article{kramer2016scikit,
   title={Scikit-learn},
-  author={Kramer, Oliver and Kramer, Oliver},
+  author={Kramer, Oliver},
   journal={Machine Learning for Evolution Strategies},
[…]
