+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article{althoff2023conform,
+@mastersthesis{althoff2023conform,
   title={Conform With the Wind},
   author={Althoff, Simon},
   school={Lund University},
@@ -184,7 +184,7 @@
   url={https://cran.r-project.org/web/packages/quantregForest/index.html}
 }
 
-@article{saporta2023statistical,
+@phdthesis{saporta2023statistical,
   title={Statistical tools for causal inference and forensic science},
   author={Saporta, Jason},
   year={2023},
