+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@mastersthesis{althoff2023conform,
+@thesis{althoff2023conform,
   title={Conform With the Wind},
   author={Althoff, Simon},
   school={Lund University},
@@ -169,7 +169,7 @@
   doi={10.1016/j.ijforecast.2021.11.001}
 }
 
-@mastersthesis{Prinzhorn2023,
+@thesis{Prinzhorn2023,
 	author={Derck Walther Eward Prinzhorn},
 	title={Benchmarking conformal prediction methods for time series regression},
 	school={University of Amsterdam},
@@ -184,7 +184,7 @@
   url={https://cran.r-project.org/web/packages/quantregForest/index.html}
 }
 
-@phdthesis{saporta2023statistical,
+@thesis{saporta2023statistical,
   title={Statistical tools for causal inference and forensic science},
   author={Saporta, Jason},
   year={2023},
