### 2023-09-11 5af0ece
Update paper
-Add DOIs
-Update references
