+++ paper/paper.bib
@@ -170,11 +170,11 @@
 }
 
 @thesis{Prinzhorn2023,
-	author={Derck Walther Eward Prinzhorn},
-	title={Benchmarking conformal prediction methods for time series regression},
-	school={University of Amsterdam},
-	year={2023},
-	type={Bachelor's Thesis}
+  author={Derck Walther Eward Prinzhorn},
+  title={Benchmarking conformal prediction methods for time series regression},
+  school={University of Amsterdam},
+  year={2023},
+  type={Bachelor's Thesis}
 }
 
 @misc{quantregforest2017,
