+++ _sources/auto_examples/plot_quantile_vs_standard_forest.rst.txt
@@ -84,7 +84,7 @@ distribution, the mean is to the right of the median.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.558 seconds)
+   **Total running time of the script:** (0 minutes 0.499 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_vs_standard_forest.py:
