+++ _sources/auto_examples/plot_quantile_weighted_vs_unweighted.rst.txt
@@ -137,7 +137,7 @@ comparison.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 11.831 seconds)
+   **Total running time of the script:** (0 minutes 13.757 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_weighted_vs_unweighted.py:
