+++ paper/paper.md
@@ -11,13 +11,13 @@ authors:
 affiliations:
  - name: Zillow Group, USA
    index: 1
-date: 1 August 2023
+date: 1 October 2023
 bibliography: paper.bib
 ---
 
 # Summary
 
-Quantile regression forests (QRF) is a non-parametric, tree-based ensemble method for estimating conditional quantiles [@meinshausen2006quantile]. It is a generalization of the random forests algorithm, a versatile ensemble learning algorithm, originally proposed in [@breiman2001random], that has proven extremely popular and useful as a general-purpose classification and regression method [@biau2016random;@hengl2018random;@wager2018estimation;@athey2019generalized]. Instead of outputting the weighted mean value of training labels for regression like random forests, QRF employ the weighted empirical distribution of training labels to obtain the predictive distribution. This enables QRF to output probabilistic predictions for regression problems, which are widely useful for constructing estimates of uncertainty [@@petropoulos2022forecasting].
+Quantile regression forests (QRF) is a non-parametric, tree-based ensemble method for estimating conditional quantiles [@meinshausen2006quantile]. It is a generalization of the random forests algorithm, a versatile ensemble learning algorithm, originally proposed in [@breiman2001random], that has proven extremely popular and useful as a general-purpose machine learning method [@biau2016random;@hengl2018random;@wager2018estimation;@athey2019generalized]. Instead of outputting the weighted mean value of training labels for regression like random forests, QRF employ the weighted empirical distribution of training labels to obtain the predictive distribution. This enables QRF to output probabilistic predictions for regression problems, which are widely useful for constructing estimates of uncertainty [@petropoulos2022forecasting].
 
 `quantile-forest` provides a fast, feature-rich implementation of the QRF algorithm. The estimators provided in this package are optimized using Cython [@behnel2010cython] for training and inference speed, and can estimate arbitrary quantiles at prediction time without retraining. They provide methods for out-of-bag estimation, calculating quantile ranks, and computing proximity counts. The provided QRF estimators are also compatible with and can serve as drop-in replacements for the widely used forest regressors available in scikit-learn [@kramer2016scikit]. The package was designed to be used by a broad array of researchers and in production business settings. It has already been cited in scholarly work [@althoff2023conform;@saporta2023statistical;@Prinzhorn2023] and used in a production setting at the real estate technology company Zillow. The combination of speed, design, and functionality in `quantile-forest` enables exciting scientific explorations in academia and industry alike.
 
@@ -25,17 +25,17 @@ Quantile regression forests (QRF) is a non-parametric, tree-based ensemble metho
 
 Quantile regression is useful for understanding relationships between variables outside of the mean of the data, which can be particularly useful for understanding outcomes that are non-normally distributed or that have nonlinear relationships with predictor variables. It can be used to understand an outcome at its various quantiles and to compare groups or levels of an exposure on those quantiles [@koenker2005quantile].
 
-QRF, an extension of the random forest algorithm, provide a flexible, nonlinear and nonparametric way of performing quantile regression on the predictive distributions for high-dimensional data. Unlike traditional machine learning algorithms that focus solely on point estimation, QRF enables researchers to obtain a more comprehensive understanding of the underlying data distribution. It provides predictions not only for the expected outcome but also for various quantiles, allowing researchers to quantify uncertainties and capture the full spectrum of potential outcomes. QRF has been a standard method for probabilistic prediction in machine learning and applied to many areas requiring reliable probabilistic predictions.
[…]
