+++ _sources/auto_examples/plot_quantile_regression_intervals.rst.txt
@@ -163,7 +163,7 @@ intervals on the California Housing dataset.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 5.601 seconds)
+   **Total running time of the script:** (0 minutes 4.611 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_regression_intervals.py:
