+++ .github/workflows/draft-pdf.yml
@@ -9,7 +9,7 @@ jobs:
 
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
