+++ _sources/auto_examples/plot_quantile_extrapolation_problem.rst.txt
@@ -127,7 +127,7 @@ approach.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.480 seconds)
+   **Total running time of the script:** (0 minutes 0.393 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_extrapolation_problem.py:
