+++ .github/workflows/publish.yml
@@ -23,7 +23,7 @@ jobs:
           python-version: '3.11'
 
       - name: Install cibuildwheel
-        run: python -m pip install cibuildwheel==2.14.1
+        run: python -m pip install cibuildwheel==2.16.2
 
       - name: Build wheels for Linux
         if: matrix.os == 'ubuntu-latest'
