### 2023-10-04 d81b591
Speeds up weighted quantile calculation
Order quantiles monotonically and reuse floor and ceil indices.
