### 2023-10-04 b7f14eb
Update _quantile_forest_fast.pyx
