### 2023-10-04 8b4beec
Speed-up Weighted Quantile Calculation (#11)
* Speeds up weighted quantile calculation

Order quantiles monotonically and reuse floor and ceil indices.
