### 2023-10-04 1cfb551
Update builds to Python 3.12 (#12)
