+++ _sources/auto_examples/plot_quantile_toy_example.rst.txt
@@ -36,8 +36,20 @@ the predictions to a ground truth function used to generate noisy samples.
    :class: sphx-glr-single-img
 
 
+.. rst-class:: sphx-glr-script-out
+
+ .. code-block:: none
+
+    /home/runner/work/quantile-forest/quantile-forest/examples/plot_quantile_toy_example.py:40: SyntaxWarning: invalid escape sequence '\,'
+      qrf = RandomForestQuantileRegressor(max_depth=3, min_samples_leaf=5, random_state=0)
 
 
+
+
+
+
+
+|
 
 .. code-block:: default
 
@@ -97,7 +109,7 @@ the predictions to a ground truth function used to generate noisy samples.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.272 seconds)
+   **Total running time of the script:** (0 minutes 0.333 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_toy_example.py:
