+++ _sources/auto_examples/plot_quantile_toy_example.rst.txt
@@ -69,7 +69,7 @@ the predictions to a ground truth function used to generate noisy samples.
         f = x * np.sin(x)
 
         sigma = 0.25 + x / 10
-        noise = rng.lognormal(sigma=sigma) - np.exp(sigma**2 / 2)
+        noise = rng.lognormal(sigma=sigma) - np.exp(sigma ** 2 / 2)
         y = f + noise
 
         return np.atleast_2d(x).T, y
@@ -109,7 +109,7 @@ the predictions to a ground truth function used to generate noisy samples.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.362 seconds)
+   **Total running time of the script:** (0 minutes 0.443 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_toy_example.py:
