### 2023-10-05 495874c
Update paper references
