+++ paper/paper.bib
@@ -7,14 +7,14 @@
 }
 
 @article{athey2019generalized,
-  title={Generalized random forests},
+  title={Generalized Random Forests},
   author={Athey, Susan and Tibshirani, Julie and Wager, Stefan},
   year={2019},
   doi={10.1214/18-aos1709}
 }
 
 @article{behnel2010cython,
-  title={Cython: The best of both worlds},
+  title={Cython: The Best of Both Worlds},
   author={Behnel, Stefan and Bradshaw, Robert and Citro, Craig and Dalcin, Lisandro and Seljebotn, Dag Sverre and Smith, Kurt},
   journal={Computing in Science \& Engineering},
   volume={13},
@@ -26,7 +26,7 @@
 }
 
 @article{biau2016random,
-  title={A random forest guided tour},
+  title={A Random Forest Guided Tour},
   author={Biau, G{\'e}rard and Scornet, Erwan},
   journal={Test},
   volume={25},
@@ -37,7 +37,7 @@
 }
 
 @article{breiman2001random,
-  title={Random forests},
+  title={Random Forests},
   author={Breiman, Leo},
   journal={Machine learning},
   volume={45},
@@ -48,7 +48,7 @@
 }
 
 @article{cordoba2021spatially,
-  title={A spatially based quantile regression forest model for mapping rural land values},
+  title={A Spatially Based Quantile Regression Forest Model for Mapping Rural Land Values},
   author={C{\'o}rdoba, Mariano and Carranza, Juan Pablo and Piumetto, Mario and Monzani, Federico and Balzarini, M{\'o}nica},
   journal={Journal of Environmental Management},
   volume={289},
@@ -59,7 +59,7 @@
 }
 
 @article{dean2022quantile,
-  title={Quantile regression forests for individualized surgery scheduling},
+  title={Quantile Regression Forests for Individualized Surgery Scheduling},
   author={Dean, Arlen and Meisami, Amirhossein and Lam, Henry and Van Oyen, Mark P and Stromblad, Christopher and Kastango, Nick},
   journal={Health Care Management Science},
   volume={25},
@@ -71,7 +71,7 @@
 }
 
 @article{fang2018quantile,
-  title={A quantile regression forest based method to predict drug response and assess prediction reliability},
+  title={A Quantile Regression Forest Based Method to Predict Drug Response and Assess Prediction Reliability},
   author={Fang, Yun and Xu, Peirong and Yang, Jialiang and Qin, Yufang},
   journal={PLoS One},
   volume={13},
@@ -82,7 +82,7 @@
 }
 
 @article{francke2008estimation,
-  title={Estimation of suspended sediment concentration and yield using linear models, random forests and quantile regression forests},
+  title={Estimation of Suspended Sediment Concentration and Yield Using Linear Models, Random Forests and Quantile Regression Forests},
   author={Francke, T and L{\'o}pez-Taraz{\'o}n, JA and Schr{\"o}der, B},
   journal={Hydrological Processes},
   volume={22},
@@ -94,7 +94,7 @@
 }
 
 @article{gyamerah2020long,
-  title={Long-term exchange rate probability density forecasting using Gaussian kernel and quantile random forest},
+  title={Long-Term Exchange Rate Probability Density Forecasting Using {G}aussian Kernel and Quantile Random Forest},
   author={Gyamerah, Samuel Asante and Moyo, Edwin},
   journal={Complexity},
   volume={2020},
@@ -105,7 +105,7 @@
 }
 
 @article{hengl2018random,
-  title={Random forest as a generic framework for predictive modeling of spatial and spatio-temporal variables},
[…]
