+++ paper/paper.md
@@ -29,7 +29,7 @@ QRF, an extension of the random forest algorithm, provides a flexible, nonlinear
 
 Traditional prediction intervals often rely on assumptions such as normality, which may not hold in many real-world scenarios [@gyamerah2020long]. QRF, on the other hand, allows researchers to generate prediction intervals that are non-parametric, flexible, and adaptive to different data distributions. This capability is invaluable for quantifying uncertainties in a wide range of research areas, including finance [@cordoba2021spatially], environmental sciences [@francke2008estimation;@fang2018quantile;@zhang2018parallel], healthcare [@molinder2020probabilistic;@dean2022quantile], and more. A crucial difference between QRF and many other quantile regression approaches is that after training a QRF once, one has access to all the quantiles at inference time, whereas most approaches require retraining separately for each quantile.
 
-As a cutting-edge statistical modeling technique, the QRF algorithm holds immense potential for researchers across many domains, providing them with a powerful tool to address complex problems involving quantile regression and uncertainty estimation. The QRF algorithm is broadly available in R, which is host to the canonical QRF implementation [@quantregforest2017] as well as established alternative implementations [@athey2019generalized]. However Python has emerged as a prevailing standard programming language within the scientific community, making it a popular option for researchers and practitioners. The absence of a comprehensive Python implementation of the QRF algorithm severely hampers researchers' ability to utilize and benefit from its wide-ranging applications.
+As a cutting-edge statistical modeling technique, the QRF algorithm holds immense potential for researchers across many domains, providing them with a powerful tool to address complex problems involving quantile regression and uncertainty estimation. The QRF algorithm is broadly available in R, which is host to the canonical QRF implementation [@quantregforest2017] as well as established alternative implementations [@wright2017ranger;@athey2019generalized]. However Python has emerged as a prevailing standard programming language within the scientific community, making it a popular option for researchers and practitioners. The absence of a comprehensive Python implementation of the QRF algorithm severely hampers researchers' ability to utilize and benefit from its wide-ranging applications.
 
 We seek to fill this need by providing a comprehensive Python-based implementation of the QRF algorithm. The QRF implementation provided in this package has been optimized for training and inference speed. It allows specifying prediction quantiles after training, permitting a trained model to be reused to estimate conditional quantiles as needed. In addition to this base prediction functionality, the package also includes utilities that enhance the algorithm's applicability and usefulness for researchers and practitioners. These utilities include:
 
