+++ paper/paper.bib
@@ -210,7 +210,7 @@
   journal={Journal of Statistical Software},
   volume={77},
   number={1},
-  pages={1–17},
+  pages={1–-17},
   year={2017},
   url={https://www.jstatsoft.org/index.php/jss/article/view/v077i01},
   doi={10.18637/jss.v077.i01}
