+++ quantile_forest/version.txt
@@ -1,1 +1,1 @@
-1.2.1dev
+1.2.2dev
