### 2023-10-07 6311fa2
Speed-up Single Leaf Scoring
Also updates tests and docs.
