+++ .github/workflows/github-pages.yml
@@ -2,8 +2,6 @@ name: docs
 
 on:
   push:
-    branches:
-      - main
     tags:
       - 'v*.*.*'
 
