### 2023-10-07 11b05c9
Update predict unit tests
