### 2023-10-07 b9a764d
Speed-up Single Leaf Scoring (#13)
* Speed-up Single Leaf Scoring

Also updates tests and docs.
