+++ examples/plot_quantile_extrapolation_problem.py
@@ -68,8 +68,8 @@ ax1.fill_between(x_mid.ravel(), y_pred_mid[:, 0], y_pred_mid[:, 2], color="#e0f2
 ax1.plot(x_mid, y_pred_mid[:, 1], c="#006aff", lw=2)
 ax1.set_xlim(bounds)
 ax1.set_ylim([-8, 8])
-ax1.set_xlabel("x")
-ax1.set_ylabel("y")
+ax1.set_xlabel("$x$")
+ax1.set_ylabel("$y$")
 ax1.set_title("Prediction Intervals on Training Data")
 
 ax2.plot(x, y, c="#f2a619", lw=0, marker=".", ms=4)
@@ -82,8 +82,8 @@ ax2.plot(xx_left, y_pred_left[:, 1], alpha=0.8, c="r", lw=3)
 ax2.plot(xx_right, y_pred_right[:, 1], alpha=0.8, c="r", lw=3)
 ax2.set_xlim(bounds)
 ax2.set_ylim([-8, 8])
-ax2.set_xlabel("x")
-ax2.set_ylabel("y")
+ax2.set_xlabel("$x$")
+ax2.set_ylabel("$y$")
 ax2.set_title("Prediction Intervals with Extrapolated Values")
 
 plt.subplots_adjust(top=0.15)
