### 2023-10-08 81b06e5
Update example plots
