+++ examples/plot_quantile_extrapolation_problem.py
@@ -21,72 +21,92 @@ np.random.seed(0)
 n_samples = 1000
 bounds = [-1, 21]
 extrap_pct = 0.2
+func = np.sin
 
-x = np.linspace(bounds[0], bounds[1], n_samples)
 
-f = np.sin(x)
-std = 0.01 + np.abs(x - 5.0) / 5.0
-noise = np.random.normal(scale=std)
-y = f + noise
+def make_func_Xy(func, bounds, n_samples):
+    x = np.linspace(bounds[0], bounds[1], n_samples)
+    f = func(x)
+    std = 0.01 + np.abs(x - 5.0) / 5.0
+    noise = np.random.normal(scale=std)
+    y = f + noise
+    return x, y
+
+
+def get_train_Xy(X, y, min_idx, max_idx):
+    X_train = X[min_idx:max_idx]
+    y_train = y[min_idx:max_idx]
+    return X_train, y_train
+
+
+def get_test_X(X):
+    n_samples = len(X)
+    X_test = np.atleast_2d(np.linspace(-1, 21, n_samples)).T
+    return X_test
+
+
+X, y = make_func_Xy(func, bounds, n_samples)
 
 extrap_min_idx = int(n_samples * (extrap_pct / 2))
 extrap_max_idx = int(n_samples - (n_samples * (extrap_pct / 2)))
 
-x_train = x[extrap_min_idx:extrap_max_idx]
-y_train = y[extrap_min_idx:extrap_max_idx]
-
-x_mid = x[extrap_min_idx:extrap_max_idx]
-x_left = x[:extrap_min_idx]
-x_right = x[extrap_max_idx:]
-
-y_mid = y[extrap_min_idx:extrap_max_idx]
-y_left = y[:extrap_min_idx]
-y_right = y[extrap_max_idx:]
-
-xx = np.atleast_2d(np.linspace(-1, 21, n_samples)).T
-xx_mid = xx[extrap_min_idx:extrap_max_idx]
-xx_left = xx[:extrap_min_idx]
-xx_right = xx[extrap_max_idx:]
+X_train, y_train = get_train_Xy(X, y, extrap_min_idx, extrap_max_idx)
+X_test = get_test_X(X)
 
 qrf = RandomForestQuantileRegressor(
     max_samples_leaf=None,
     min_samples_leaf=10,
     random_state=0,
 )
-qrf.fit(np.expand_dims(x_train, axis=-1), y_train)
+qrf.fit(np.expand_dims(X_train, axis=-1), y_train)
 
-y_pred = qrf.predict(xx, quantiles=[0.025, 0.5, 0.975])  # extrapolate
-y_pred_mid = y_pred[extrap_min_idx:extrap_max_idx]
-y_pred_left = y_pred[:extrap_min_idx]
-y_pred_right = y_pred[extrap_max_idx:]
+y_pred = qrf.predict(X_test, quantiles=[0.025, 0.5, 0.975])  # extrapolate
 
-fig, (ax1, ax2) = plt.subplots(nrows=1, ncols=2, figsize=(10, 4))
 
-ax1.plot(x_train, y_train, c="#f2a619", lw=0, marker=".", ms=4)
-ax1.plot(x_train, f[extrap_min_idx:extrap_max_idx], c="black")
-ax1.fill_between(x_mid.ravel(), y_pred_mid[:, 0], y_pred_mid[:, 2], color="#e0f2ff")
-ax1.plot(x_mid, y_pred_mid[:, 1], c="#006aff", lw=2)
-ax1.set_xlim(bounds)
-ax1.set_ylim([-8, 8])
[…]
