+++ examples/plot_quantile_regression_intervals.py
@@ -20,8 +20,6 @@ from quantile_forest import RandomForestQuantileRegressor
 
 rng = check_random_state(0)
 
-dollar_formatter = FuncFormatter(lambda x, p: "$" + format(int(x), ","))
-
 # Load the California Housing Prices dataset.
 california = datasets.fetch_california_housing()
 n_samples = min(california.target.size, 1000)
@@ -36,8 +34,8 @@ kf.get_n_splits(X)
 
 y_true = []
 y_pred = []
-y_pred_lower = []
-y_pred_upper = []
+y_pred_low = []
+y_pred_upp = []
 
 for train_index, test_index in kf.split(X):
     X_train, X_test, y_train, y_test = (
@@ -55,74 +53,79 @@ for train_index, test_index in kf.split(X):
 
     y_true = np.concatenate((y_true, y_test))
     y_pred = np.concatenate((y_pred, y_pred_i[:, 1]))
-    y_pred_lower = np.concatenate((y_pred_lower, y_pred_i[:, 0]))
-    y_pred_upper = np.concatenate((y_pred_upper, y_pred_i[:, 2]))
+    y_pred_low = np.concatenate((y_pred_low, y_pred_i[:, 0]))
+    y_pred_upp = np.concatenate((y_pred_upp, y_pred_i[:, 2]))
 
-# Scale data to dollars.
-y_true *= 1e5
-y_pred *= 1e5
-y_pred_lower *= 1e5
-y_pred_upper *= 1e5
 
-fig, (ax1, ax2) = plt.subplots(nrows=1, ncols=2, figsize=(10, 4))
+def plot_calibration_and_intervals(y_true, y_pred, y_pred_low, y_pred_upp):
+    def plot_calibration(ax, y_true, y_pred_low, y_pred_upp, price_formatter):
+        y_min = min(np.minimum(y_true, y_pred))
+        y_max = max(np.maximum(y_true, y_pred))
 
-y_pred_interval = y_pred_upper - y_pred_lower
-sort_idx = np.argsort(y_pred)
-y_true = y_true[sort_idx]
-y_pred = y_pred[sort_idx]
-y_pred_lower = y_pred_lower[sort_idx]
-y_pred_upper = y_pred_upper[sort_idx]
-y_min = min(np.minimum(y_true, y_pred))
-y_max = max(np.maximum(y_true, y_pred))
-y_min = float(np.round((y_min / 10000) - 1, 0) * 10000)
-y_max = float(np.round((y_max / 10000) - 1, 0) * 10000)
+        for low, mid, upp in zip(y_pred_low, y_pred, y_pred_upp):
+            ax.plot([mid, mid], [low, upp], lw=4, c="#e0f2ff")
 
-for low, mid, upp in zip(y_pred_lower, y_pred, y_pred_upper):
-    ax1.plot([mid, mid], [low, upp], lw=4, c="#e0f2ff")
+        ax.plot(y_pred, y_true, c="#f2a619", lw=0, marker=".", ms=5)
+        ax.plot(y_pred, y_pred_low, alpha=0.4, c="#006aff", lw=0, marker="_", ms=4)
+        ax.plot(y_pred, y_pred_upp, alpha=0.4, c="#006aff", lw=0, marker="_", ms=4)
+        ax.plot([y_min, y_max], [y_min, y_max], ls="--", lw=1, c="grey")
+        ax.grid(axis="x", color="0.95")
+        ax.grid(axis="y", color="0.95")
+        ax.xaxis.set_major_formatter(price_formatter)
+        ax.yaxis.set_major_formatter(price_formatter)
+        ax.set_xlim(y_min, y_max)
[…]
