### 2023-10-08 013cf8b
Add mapie to sphinx_requirements
