+++ docs/sphinx_requirements.txt
@@ -2,7 +2,7 @@ sphinx
 sphinx_gallery
 sphinx_rtd_theme
 sphinxcontrib.bibtex
+pydata_sphinx_theme
 mapie
 numpydoc
 matplotlib
-pydata_sphinx_theme
