### 2023-10-08 5e72bd8
Update Docs to PyData Theme (#14)
* Updates docs theme
* Updates example plots
* Adds plot on CQR
