+++ examples/plot_quantile_toy_example.py
@@ -42,21 +42,26 @@ qrf.fit(X_train, y_train)
 
 y_pred = qrf.predict(xx, quantiles=[0.025, 0.5, 0.975])
 
-y_pred_low = y_pred[:, 0]
-y_pred_med = y_pred[:, 1]
-y_pred_upp = y_pred[:, 2]
 
-plt.plot(X_test, y_test, ".", c="#f2a619", label="Test Observations", ms=5)
-plt.plot(xx, (xx * np.sin(xx)), c="black", label="$f(x) = x\\,\\sin(x)$", lw=2)
-plt.plot(xx, y_pred_med, c="#006aff", label="Predicted Median", lw=3, ms=5)
-plt.fill_between(
-    xx.ravel(),
-    y_pred_low,
-    y_pred_upp,
-    color="#e0f2ff",
-    label="Predicted 95% Interval",
-)
-plt.xlabel("$x$")
-plt.ylabel("$f(x)$")
-plt.legend(loc="upper left")
-plt.show()
+def plot_fit(X_test, y_test, y_pred):
+    y_pred_low = y_pred[:, 0]
+    y_pred_med = y_pred[:, 1]
+    y_pred_upp = y_pred[:, 2]
+
+    plt.plot(X_test, y_test, ".", c="#f2a619", label="Test Observations", ms=5)
+    plt.plot(xx, (xx * np.sin(xx)), c="black", label="$f(x) = x\\,\\sin(x)$", lw=2)
+    plt.plot(xx, y_pred_med, c="#006aff", label="Predicted Median", lw=3, ms=5)
+    plt.fill_between(
+        xx.ravel(),
+        y_pred_low,
+        y_pred_upp,
+        color="#e0f2ff",
+        label="Predicted 95% Interval",
+    )
+    plt.xlabel("$x$")
+    plt.ylabel("$f(x)$")
+    plt.legend(loc="upper left")
+    plt.show()
+
+
+plot_fit(X_test, y_test, y_pred)
