+++ examples/plot_quantile_conformalized.py
@@ -1,0 +1,198 @@
+"""
+========================================================================
+Quantile regression forests for conformalized quantile regression (CQR)
+========================================================================
+
+An example that demonstrates the use of a quantile regression forest to
+construct reliable prediction intervals using conformalized quantile
+regression (CQR). CQR offers prediction intervals that attain valid coverage,
+while QRF may require additional calibration for reliable interval estimates.
+This example uses MAPIE to construct the CQR interval estimates.
+"""
+print(__doc__)
+
+import warnings
+
+import matplotlib.pyplot as plt
+import numpy as np
+from matplotlib.offsetbox import AnchoredText
+from matplotlib.ticker import FuncFormatter
+from mapie.metrics import regression_coverage_score, regression_mean_width_score
+from mapie.regression import MapieQuantileRegressor
+from sklearn import datasets
+from sklearn.model_selection import train_test_split
+from sklearn.utils.validation import check_random_state
+from quantile_forest import RandomForestQuantileRegressor
+
+random_state = 0
+rng = check_random_state(0)
+round_to = 3
+usd_formatter = FuncFormatter(lambda x, p: f"${format(int(x) * 100, ',')}k")
+alpha = 0.5
+
+# Load the California Housing Prices dataset.
+california = datasets.fetch_california_housing()
+n_samples = min(california.target.size, 1000)
+perm = rng.permutation(n_samples)
+X = california.data[perm]
+y = california.target[perm]
+
+X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=random_state)
+X_train, X_calib, y_train, y_calib = train_test_split(
+    X_train, y_train, test_size=0.5, random_state=random_state
+)
+
+
+class WrappedRandomForestQuantileRegressor(RandomForestQuantileRegressor):
+    """Wrap the QRF estimator with the parameters expected by MAPIE."""
+
+    def __init__(self, loss="quantile", **kwargs):
+        super().__init__(**kwargs)
+        self.init_dict = {"loss": loss}
+        self.loss = loss
+        self.kwargs = kwargs
+
+    @classmethod
+    def parent(cls):
+        return cls.__bases__[0]
+
+    def get_params(self, *args, **kwargs):
+        params = self.init_dict
+        params.update(self.parent()(**self.kwargs).get_params(*args, **kwargs))
+        return params
+
+
+def sort_y_values(y_test, y_pred, y_pis):
+    """Sort the dataset in order to make plots using the `fill_between` function."""
+    indices = np.argsort(y_test)
+    y_test_sorted = np.array(y_test)[indices]
+    y_pred_sorted = y_pred[indices]
+    y_lower_bound = y_pis[:, 0, 0][indices]
+    y_upper_bound = y_pis[:, 1, 0][indices]
+    return y_test_sorted, y_pred_sorted, y_lower_bound, y_upper_bound
+
+
+def plot_prediction_intervals(
+    title,
+    ax,
+    y_test,
+    y_pred,
+    y_pred_low,
+    y_pred_upp,
+    coverage,
+    width,
+    num_plots_idx,
+    price_formatter,
+):
+    """Plot of the prediction intervals for each different conformal method."""
+    y_pred_low_ = np.take(y_pred_low, num_plots_idx)
+    y_pred_upp_ = np.take(y_pred_upp, num_plots_idx)
[…]
