### 2023-10-08 b1f8cba
Refactor example
