### 2023-10-08 e76d407
Refactor CQR example
