### 2023-10-08 267b492
Update CQR example
