### 2023-10-08 049d7a8
Refactor examples
