### 2023-10-08 9d9db12
Update and refactor examples
