+++ _downloads/9692ea47a09c86e0bd5b0eaab224f0dd/plot_quantile_interpolation.py
@@ -14,7 +14,6 @@ import numpy as np
 
 from quantile_forest import RandomForestQuantileRegressor
 
-
 # Create toy dataset.
 X = np.array([[-1, -1], [-1, -1], [-1, -1], [1, 1], [1, 1]])
 y = np.array([-2, -1, 0, 1, 2])
