+++ _sources/auto_examples/plot_quantile_interpolation.rst.txt
@@ -25,7 +25,7 @@ Predicting with different quantile interpolation methods
 An example comparison of interpolation methods that can be applied during
 prediction when the desired quantile lies between two data points.
 
-.. GENERATED FROM PYTHON SOURCE LINES 10-77
+.. GENERATED FROM PYTHON SOURCE LINES 10-76
 
 
 
@@ -46,7 +46,6 @@ prediction when the desired quantile lies between two data points.
     import numpy as np
 
     from quantile_forest import RandomForestQuantileRegressor
-
 
     # Create toy dataset.
     X = np.array([[-1, -1], [-1, -1], [-1, -1], [1, 1], [1, 1]])
@@ -111,7 +110,7 @@ prediction when the desired quantile lies between two data points.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.108 seconds)
+   **Total running time of the script:** (0 minutes 0.109 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_interpolation.py:
