+++ _sources/auto_examples/plot_quantile_regression_intervals.rst.txt
@@ -169,7 +169,7 @@ intervals on the California Housing dataset.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 4.692 seconds)
+   **Total running time of the script:** (0 minutes 4.696 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_regression_intervals.py:
