### 2023-10-08 a676969
Add unit test for interpolation
