+++ quantile_forest/version.txt
@@ -1,1 +1,1 @@
-1.2.2dev
+1.2.3
