### 2023-10-09 5077601
Update CQR example text
