+++ _sources/auto_examples/plot_quantile_conformalized.rst.txt
@@ -18,11 +18,11 @@
 .. _sphx_glr_auto_examples_plot_quantile_conformalized.py:
 
 
-========================================================================
-Quantile regression forests for conformalized quantile regression (CQR)
-========================================================================
+=================================================================
+Quantile regression forests for conformalized quantile regression
+=================================================================
 
-An example that demonstrates the use of a quantile regression forest to
+An example that demonstrates the use of a quantile regression forest (QRF) to
 construct reliable prediction intervals using conformalized quantile
 regression (CQR). CQR offers prediction intervals that attain valid coverage,
 while QRF may require additional calibration for reliable interval estimates.
@@ -246,7 +246,7 @@ This example uses MAPIE to construct the CQR interval estimates with a QRF.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 2.469 seconds)
+   **Total running time of the script:** (0 minutes 2.969 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_conformalized.py:
