+++ _sources/auto_examples/plot_quantile_conformalized.rst.txt
@@ -246,7 +246,7 @@ This example uses MAPIE to construct the CQR interval estimates with a QRF.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 2.969 seconds)
+   **Total running time of the script:** (0 minutes 3.034 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_conformalized.py:
