### 2023-10-10 4f87d96
Add input checks
