+++ _downloads/cf7a0c743e22634b5b6a152569efba36/plot_quantile_conformalized.py
@@ -187,7 +187,7 @@ fig, axs = plt.subplots(nrows=1, ncols=2, figsize=(10, 4.15))
 
 coords = [axs[0], axs[1]]
 num_plots = rng.choice(len(y_test), int(len(y_test)), replace=False)
-usd_formatter = FuncFormatter(lambda x, p: f"${format(int(x) * 100, ',')}k")
+usd_formatter = FuncFormatter(lambda x, p: f"${format(int(x * 100), ',')}k")
 
 for strategy, coord in zip(strategies.keys(), coords):
     plot_prediction_intervals(
