+++ _sources/auto_examples/plot_quantile_toy_example.rst.txt
@@ -103,7 +103,7 @@ the predictions to a ground truth function used to generate noisy samples.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.427 seconds)
+   **Total running time of the script:** (0 minutes 0.466 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_toy_example.py:
