+++ _sources/auto_examples/plot_quantile_conformalized.rst.txt
@@ -221,7 +221,7 @@ This example uses MAPIE to construct the CQR interval estimates with a QRF.
 
     coords = [axs[0], axs[1]]
     num_plots = rng.choice(len(y_test), int(len(y_test)), replace=False)
-    usd_formatter = FuncFormatter(lambda x, p: f"${format(int(x) * 100, ',')}k")
+    usd_formatter = FuncFormatter(lambda x, p: f"${format(int(x * 100), ',')}k")
 
     for strategy, coord in zip(strategies.keys(), coords):
         plot_prediction_intervals(
@@ -246,7 +246,7 @@ This example uses MAPIE to construct the CQR interval estimates with a QRF.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 3.034 seconds)
+   **Total running time of the script:** (0 minutes 3.118 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_conformalized.py:
