### 2023-10-11 ad0d421
F-strings and linting
