+++ quantile_forest/tests/test_quantile_forest.py
@@ -4,7 +4,7 @@ Testing for the quantile forest module (quantile_forest._quantile_forest).
 
 import math
 import warnings
-from typing import Any, Dict
+from typing import Any
 
 import numpy as np
 import pytest
@@ -41,7 +41,7 @@ perm = rng.permutation(min(california.target.size, 500))
 X_california = california.data[perm]
 y_california = california.target[perm]
 
-FOREST_REGRESSORS: Dict[str, Any] = {
+FOREST_REGRESSORS: dict[str, Any] = {
     "ExtraTreesQuantileRegressor": ExtraTreesQuantileRegressor,
     "RandomForestQuantileRegressor": RandomForestQuantileRegressor,
 }
@@ -58,7 +58,7 @@ def check_regression_toy(name, weighted_quantile):
 
     ForestRegressor = FOREST_REGRESSORS[name]
 
-    regr = ForestRegressor(n_estimators=10, max_samples_leaf=None, bootstrap=False, random_state=1)
+    regr = ForestRegressor(n_estimators=10, max_samples_leaf=None, bootstrap=False, random_state=0)
     regr.fit(X, y)
 
     # Check model and apply outputs shape.
@@ -97,30 +97,23 @@ def check_california_criterion(name, criterion):
     # Check for consistency on the California Housing Prices dataset.
     ForestRegressor = FOREST_REGRESSORS[name]
 
-    regr = ForestRegressor(n_estimators=5, criterion=criterion, random_state=1)
+    regr = ForestRegressor(n_estimators=5, criterion=criterion, max_features=None, random_state=0)
     regr.fit(X_california, y_california)
     score = regr.score(X_california, y_california, quantiles=0.5)
-    assert score > 0.85, "Failed with max_features=None, criterion {0} and score = {1}".format(
-        criterion, score
-    )
+    assert score > 0.9, f"Failed with max_features=None, criterion {criterion} and score={score}."
 
     # Test maximum features.
-    regr = ForestRegressor(n_estimators=5, criterion=criterion, max_features=6, random_state=1)
+    regr = ForestRegressor(n_estimators=5, criterion=criterion, max_features=6, random_state=0)
     regr.fit(X_california, y_california)
     score = regr.score(X_california, y_california, quantiles=0.5)
-    assert score > 0.9, "Failed with max_features=6, criterion {0} and score = {1}".format(
-        criterion, score
-    )
+    assert score > 0.9, f"Failed with max_features=6, criterion {criterion} and score={score}."
 
     # Test sample weights.
-    regr = ForestRegressor(n_estimators=5, criterion=criterion, random_state=1)
-
+    regr = ForestRegressor(n_estimators=5, criterion=criterion, random_state=0)
     sample_weight = np.ones(y_california.shape)
     regr.fit(X_california, y_california, sample_weight=sample_weight)
     score = regr.score(X_california, y_california, quantiles=0.5)
-    assert score > 0.9, "Failed with criterion {0}, sample weight and score = {1}".format(
-        criterion, score
-    )
+    assert score > 0.9, f"Failed with criterion {criterion}, sample weight and score={score}."
 
 
 @pytest.mark.parametrize("name", FOREST_REGRESSORS)
@@ -951,7 +944,13 @@ def test_predict_oob(
     weighted_quantile,
     aggregate_leaves_first,
 ):
-    check_predict_oob(name, max_samples_leaf, quantiles, weighted_quantile, aggregate_leaves_first)
+    check_predict_oob(
+        name,
+        max_samples_leaf,
+        quantiles,
+        weighted_quantile,
+        aggregate_leaves_first,
+    )
 
 
[…]
