### 2023-10-11 c465384
Fix Cython warnings
