### 2023-10-11 405ed43
Use sklearn param validation
