+++ quantile_forest/_quantile_forest.py
@@ -24,6 +24,7 @@ Only single output problems are handled.
 
 import random
 import warnings
+from abc import abstractmethod
 from math import ceil
 from numbers import Integral, Real
 from warnings import warn
@@ -39,6 +40,12 @@ from sklearn.ensemble._forest import (
 from sklearn.tree import DecisionTreeRegressor, ExtraTreeRegressor
 from sklearn.tree._tree import DTYPE
 from sklearn.utils import parse_version
+
+param_validation = True
+try:
+    from sklearn.utils._param_validation import Interval, RealNotInt
+except ImportError:
+    param_validation = False
 from sklearn.utils.validation import check_is_fitted
 
 from ._quantile_forest_fast import QuantileForest, generate_unsampled_indices
@@ -72,6 +79,50 @@ class BaseForestQuantileRegressor(ForestRegressor):
     instead.
     """
 
+    if param_validation:
+        _parameter_constraints: dict = {
+            **ForestRegressor._parameter_constraints,
+            **DecisionTreeRegressor._parameter_constraints,
+            "max_samples_leaf": [
+                None,
+                Interval(RealNotInt, 0, 1, closed="right"),
+                Interval(Integral, 1, None, closed="left"),
+            ],
+        }
+        _parameter_constraints.pop("splitter")
+
+    @abstractmethod
+    def __init__(
+        self,
+        estimator,
+        n_estimators=100,
+        *,
+        estimator_params=tuple(),
+        bootstrap=False,
+        oob_score=False,
+        n_jobs=None,
+        random_state=None,
+        verbose=0,
+        warm_start=False,
+        max_samples=None,
+        max_samples_leaf=1,
+    ):
+        init_dict = {
+            (
+                "base_estimator" if sklearn_version < parse_version("1.2.0") else "estimator"
+            ): estimator,
+            "n_estimators": n_estimators,
+            "estimator_params": estimator_params,
+            "bootstrap": bootstrap,
+            "oob_score": oob_score,
+            "n_jobs": n_jobs,
+            "random_state": random_state,
+            "verbose": verbose,
+            "warm_start": warm_start,
+            "max_samples": max_samples,
+        }
+        super().__init__(**init_dict)
+
     def fit(self, X, y, sample_weight=None, sparse_pickle=False):
         """Build a forest from the training set (X, y).
 
@@ -100,6 +151,26 @@ class BaseForestQuantileRegressor(ForestRegressor):
         self : object
             Fitted estimator.
         """
+        if param_validation:
+            self._validate_params()
+        else:
+            if isinstance(self.max_samples_leaf, (Integral, np.integer)):
+                if self.max_samples_leaf < 1:
+                    raise ValueError(
+                        "If max_samples_leaf is an integer, it must be be >= 1, "
+                        f"got {self.max_samples_leaf}."
+                    )
+            elif isinstance(self.max_samples_leaf, Real):
+                if not 0.0 < self.max_samples_leaf <= 1.0:
+                    raise ValueError(
+                        "If max_samples_leaf is a float, it must be in range (0, 1], "
+                        f"got {self.max_samples_leaf}."
+                    )
+            elif self.max_samples_leaf is not None:
+                raise ValueError(
+                    "max_samples_leaf must be of integer, float, or None type, got "
+                    f"{self.max_samples_leaf}."
+                )
         X, y = self._validate_data(X, y, multi_output=False, accept_sparse="csc", dtype=DTYPE)
         super(BaseForestQuantileRegressor, self).fit(X, y, sample_weight=sample_weight)
 
@@ -160,19 +231,9 @@ class BaseForestQuantileRegressor(ForestRegressor):
         n_samples = X.shape[0]
[…]
