### 2023-10-12 #5795 @jbytecode
@arfon - Can we trigger the `AI-based automatic reviewer recommendations with similarities` command here? I think it is automatically called in recently opened issues but this issue has been created before the tool was implemented.
