### 2023-10-13 #5795 @arfon
@jbytecode – we can. But for now, it happens after the pdf generation (I'll break it out into its own step/command :soon:)
