+++ _sources/auto_examples/plot_quantile_interpolation.rst.txt
@@ -110,7 +110,7 @@ prediction when the desired quantile lies between two data points.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.160 seconds)
+   **Total running time of the script:** (0 minutes 0.155 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_interpolation.py:
