### 2023-10-16 #5795 @ArkajyotiSaha
Hello!
Thanks so much for the invitation.
Unfortunately, I will have to decline, since I have very limited experience
with Python.

-Arka

On Mon, Oct 16, 2023 at 12:32 AM Mehmet Hakan Satman <
***@***.***> wrote:

> 👋👋👋 Dear @kpolimis <https://github.com/kpolimis>, @salrm8
> <https://github.com/salrm8>, @zmjones <https://github.com/zmjones>,
> @ArkajyotiSaha <https://github.com/ArkajyotiSaha>, @bcjaeger
> <https://github.com/bcjaeger> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled:
>
> *quantile-forest: A Python Package for Quantile Regression Forests*
>
> You can find more information at the top of this Github issue (#5795
> <https://github.com/openjournals/joss-reviews/issues/5795>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be 23 check items for each single reviewer.
>
> Thank you in advance!
>
