### 2023-10-23 #5795 @reidjohnson
@jbytecode Here are some suggestions based on a cursory review of their familiarity with Python, ML, uncertainty estimation, and/or tree-based models:

> astrogilda
> oparisot
> jncraton
> xiaowuc2
> brunaw
> YehorYudinIPP
> shahmoradi
