### 2023-10-23 #5976 @editorialbot
[REVIEW]: quantile-forest: A Python Package for Quantile Regression Forests
**Submitting author:** @reidjohnson (<a href="http://orcid.org/0009-0001-1449-4940">Reid A Johnson</a>)
**Repository:** https://github.com/zillow/quantile-forest
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.4
