### 2023-10-23 #5795 @astrogilda
I can review it as well, as long as the due date is in December, post
NeurIPS. Prior commitments.

On Mon, Oct 23, 2023, 7:56 AM Mehmet Hakan Satman ***@***.***>
wrote:

> 👋👋👋 Dear @astrogilda <https://github.com/astrogilda>, @oparisot
> <https://github.com/oparisot> , @jncraton <https://github.com/jncraton>
> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled:
>
> *quantile-forest: A Python Package for Quantile Regression Forests*
>
> You can find more information at the top of this Github issue (#5795
> <https://github.com/openjournals/joss-reviews/issues/5795>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be 23 check items for each single reviewer.
>
> Thank you in advance!
>
