### 2023-10-23 #5976 @jbytecode
@jncraton, @salrm8, @astrogilda

Thank you for accepting our invitation to review for JOSS.

This is the review thread. Firstly, type

to generate your own checklist. In that checklist, there are many check items. Whenever you complete the corresponding task, you can check them off.

Please write your comments as separate posts and do not modify your checklist descriptions.

The review process is interactive so you can always interact with the authors, reviewers, and the editor. You can also create issues and pull requests in the target repository. Please do mention this thread's URL in the issues so we can keep tracking what is going on out of our world.

Please do not hesitate to ask me about anything at anytime.

Thank you in advance!
