### 2023-10-24 zillow/quantile-forest#16 @jncraton
Unable to successfully run tests using Python 3.8
I'm attempting to test this project on all supported Python versions, which appears to include Python 3.8 to 3.12. However, I'm getting errors testing on Python 3.8. I believe that this is related to [type hinting syntax that is not supported in 3.8](https://stackoverflow.com/a/63460173).

Here's the error that I'm seeing:

    quantile_forest/tests/test_quantile_forest.py:44: in <module>
    FOREST_REGRESSORS: dict[str, Any] = {
    E   TypeError: 'type' object is not subscriptable

If it's helpful, I'm seeing this error during an automated testing run in a fork of this project. The exact error message and environment I'm using can be seen in the [Actions log](https://github.com/jncraton/quantile-forest/actions/runs/6618064069/job/17975703965).

Is Python 3.8 still supported? Are the tests expected to run correctly for that version?

This issue is related to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/5976) of this project.
