+++ examples/plot_quantile_weighted_vs_unweighted.py
@@ -9,6 +9,7 @@ predicted output values. A standard random forest regressor is included for
 comparison.
 
 """
+
 print(__doc__)
 
 import time
