### 2023-10-25 efd31f1
Update build to test Python 3.8
