### 2023-10-25 37b3c1a
Fix type hints for Python 3.8
