+++ quantile_forest/tests/test_quantile_forest.py
@@ -4,7 +4,7 @@ Testing for the quantile forest module (quantile_forest._quantile_forest).
 
 import math
 import warnings
-from typing import Any
+from typing import Any, Dict
 
 import numpy as np
 import pytest
@@ -41,7 +41,7 @@ perm = rng.permutation(min(california.target.size, 500))
 X_california = california.data[perm]
 y_california = california.target[perm]
 
-FOREST_REGRESSORS: dict[str, Any] = {
+FOREST_REGRESSORS: Dict[str, Any] = {
     "ExtraTreesQuantileRegressor": ExtraTreesQuantileRegressor,
     "RandomForestQuantileRegressor": RandomForestQuantileRegressor,
 }
