### 2023-11-09 #5976 @jncraton
@reidjohnson I have a question about how this compares to other Python packages available. In the paper, you mention other implementations in C++ and R, but I didn't see comparisons to other Python packages. Here's a quote from the paper:

> The absence of a comprehensive Python implementation of the QRF algorithm severely hampers researchers’ ability to utilize and benefit from its wide-ranging applications. We seek to fill this need by providing a comprehensive Python-based implementation of the QRF algorithm. The QRF implementation provided in this package has been optimized for training and inference speed

There may be benefit in mentioning that other packages exist and explicitly stating how your package adds value in terms of factors like performance, maintenance status, robustness, ease of use, etc.

I'm not an expert in this area, but a quick search turned up the following Python packages that also appear to offer similar functionality:

1. https://pypi.org/project/scikit-garden/
2. https://pypi.org/project/sklearn-quantile/
