### 2023-11-14 #5976 @reidjohnson
@jncraton Thanks for the suggestions here. You're correct to identify these packages as viable alternatives, and I'll update the paper to reference them accordingly. As it stands, the primary differentiators between this package and those are:

- Maintenance: One of the packages (scikit-garden) is presently unmaintained in incompatible with recent versions of Cython and scikit-learn. (Nit: This package is also presently the only one to provide prebuilt wheels for Python 3.12.)
- Performance: On large-scale data, the runtime gap between this package and the others becomes orders of magnitude. For example, fitting and predicting on the the toy dataset `X, y = datasets.make_regression(n_samples=1_000_000, n_features=10)` with comparable parameterizations takes about 3 hours for this package on a standard Macbook, but at least several days for the other packages. I don't necessarily want to do a performance bake-off, but I can provide a more thorough runtime comparison if desired or deemed useful.
- Convenience: There are some crucial differences in the way arguments are specified in the packages, and other relevant functionality. For example, in this package, quantiles can be specified at prediction time (in contrast to sklearn-quantile), and this package provides other convenience methods like out-of-bag scoring, quantile rank calculations, and proximity/similarity estimation.

So, given all that, I'm happy to iterate on the relevant parts of the paper until there's agreement. Here's an initial proposed revision of the paragraph:
> We seek to fill this need by providing a comprehensive Python-based implementation of the QRF algorithm. While other Python-based QRF implementations exist and are meaningful additions to the Python ecosystem [@scikitgarden2017;@sklearnquantile2022], none currently provide performance and functionality comparable to the implementations available in R, such as specifying quantiles at inference time or scaling to large datasets without approximation. By contrast, the QRF implementation provided in this package has been optimized for training and inference speed, enabling it to scale to millions of samples. It also allows specifying prediction quantiles after training, permitting a trained model to be reused to estimate conditional quantiles as needed. Beyond this, the package includes utilities that enhance the algorithm's applicability and usefulness for researchers and practitioners. These utilities include:
