+++ paper/paper.md
@@ -31,7 +31,7 @@ Traditional prediction intervals often rely on assumptions such as normality, wh
 
 As a cutting-edge statistical modeling technique, the QRF algorithm holds immense potential for researchers across many domains, providing them with a powerful tool to address complex problems involving quantile regression and uncertainty estimation. The QRF algorithm is broadly available in R, which is host to the canonical QRF implementation [@quantregforest2017] as well as established alternative implementations [@wright2017ranger;@athey2019generalized]. However Python has emerged as a prevailing standard programming language within the scientific community, making it a popular option for researchers and practitioners. The absence of a comprehensive Python implementation of the QRF algorithm severely hampers researchers' ability to utilize and benefit from its wide-ranging applications.
 
-We seek to fill this need by providing a comprehensive Python-based implementation of the QRF algorithm. The QRF implementation provided in this package has been optimized for training and inference speed. It allows specifying prediction quantiles after training, permitting a trained model to be reused to estimate conditional quantiles as needed. In addition to this base prediction functionality, the package also includes utilities that enhance the algorithm's applicability and usefulness for researchers and practitioners. These utilities include:
+We seek to fill this need by providing a comprehensive Python-based implementation of the QRF algorithm. While other Python-based QRF implementations exist and are meaningful additions to the Python ecosystem [@scikitgarden2017;@sklearnquantile2022], none currently provide performance and functionality comparable to the implementations available in R, such as specifying quantiles at inference time or scaling to large datasets without approximation. By contrast, the QRF implementation provided in this package has been optimized for training and inference speed, enabling it to scale to millions of samples with a runtime orders of magnitude faster than less-optimized solutions. It also allows specifying prediction quantiles after training, permitting a trained model to be reused to estimate conditional quantiles as needed. Beyond this, the package includes utilities that enhance the algorithm's applicability and usefulness for researchers and practitioners. These utilities include:
 
 * Out-of-bag scoring: The QRF algorithm can utilize the out-of-bag (OOB) samples, which are the data points that are not used during the construction of a specific decision tree in the forest. OOB scoring can be used to obtain unbiased estimates of prediction errors and quantile-specific metrics, enabling researchers to assess the performance and reliability of the QRF model without the need for additional validation datasets.
 * Quantile rank calculation: The QRF algorithm can be leveraged to facilitate the calculation of quantile ranks. Quantile ranks provide a measure of relative standing for each data point in the distribution. This capability allows researchers to compare and rank observations based on their position within the quantile distribution, providing valuable insights for various applications, such as risk assessment and anomaly detection.
