+++ CONTRIBUTING.md
@@ -2,9 +2,37 @@
 
 Contributions are welcome, encouraged, and appreciated!
 
-If you encounter any bugs while using `quantile-forest`, or believe there's a feature that would prove useful, feel free to [submit a new issue](https://github.com/zillow/quantile-forest/issues/new/choose).
+If you encounter any bugs while using the project, or believe there's a feature that would prove useful, feel free to [submit a new issue](https://github.com/zillow/quantile-forest/issues/new/choose).
 
 All contributions, suggestions, and feedback you submitted are accepted under the [project's license](https://github.com/zillow/quantile-forest/blob/main/LICENSE).
+
+## Submitting an Issue
+
+[Issues](https://github.com/zillow/quantile-forest/issues) should be used to report problems with the package or any of its dependencies, request a new feature, or to discuss potential changes before a PR is created. 
+
+When reporting bugs, please provide a [minimal reproducible example](https://stackoverflow.com/help/minimal-reproducible-example), the version of the package, and the environment (e.g., Python version).
+
+If you find an existing Issue that addresses the problem you're having, please add your own reproducible example to the existing issue rather than creating a new one. 
+
+## Submitting a Pull Request
+
+[PRs](https://github.com/zillow/quantile-forest/pulls) are always welcome and can be a quick way to get your fix or improvement merged. In general, PRs should:
+
+- Only fix/add the functionality in question.
+- Address a single concern in the fewest number of changes possible.
+- Include updated documentation.
+
+For changes that address core functionality or would require breaking changes (e.g., a major release), it's generally best to first open an Issue to discuss your proposed changes.
+
+In general, we follow the ["fork-and-pull" Git workflow](https://gist.github.com/Chaser324/ce0505fbed06b947d962)
+
+- [Fork](https://github.com/zillow/quantile-forest/fork) the repository to your own GitHub account
+- Clone the project to your machine
+- Create a branch locally with a succinct but descriptive name
+- Commit changes to the branch
+- Following any formatting and testing guidelines specific to this repo
+- Push changes to your fork
+- Open a PR in our repository
 
 ## Setting Up Your Environment
 
