+++ quantile_forest/tests/test_quantile_forest.py
@@ -609,9 +609,9 @@ def check_proximity_counts(name):
     n_samples = len(X_california)
     proximities = est.proximity_counts(X_california)
     X_leaves = est.apply(X_california)
-    bootstrap_indices = np.array(
-        [_generate_sample_indices(e.random_state, n_samples, n_samples) for e in est.estimators_]
-    )
+    bootstrap_indices = np.array([
+        _generate_sample_indices(e.random_state, n_samples, n_samples) for e in est.estimators_
+    ])
     for train_idx, train_idx_prox in enumerate(proximities):
         for proximity_idx, proximity_count in train_idx_prox:
             bootstrap_count = 0
@@ -712,16 +712,14 @@ def test_oob_samples(name):
 
 def check_oob_samples_duplicates(name):
     # Check OOB sampling with duplicates.
-    X = np.array(
-        [
-            [1, 2, 3],
-            [4, 5, 6],
-            [7, 8, 9],
-            [10, 11, 12],
-            [13, 14, 15],
-            [13, 14, 15],
-        ]
-    )
+    X = np.array([
+        [1, 2, 3],
+        [4, 5, 6],
+        [7, 8, 9],
+        [10, 11, 12],
+        [13, 14, 15],
+        [13, 14, 15],
+    ])
     y = np.array([-1, 10, 20, 30, 40, 41], dtype=np.float64)
 
     ForestRegressor = FOREST_REGRESSORS[name]
@@ -1031,9 +1029,9 @@ def check_proximity_counts_oob(name):
     # Check that OOB proximity counts match OOB bootstrap counts.
     X_leaves = est.apply(X)
     n_samples = len(X)
-    bootstrap_indices = np.array(
-        [_generate_sample_indices(e.random_state, n_samples, n_samples) for e in est.estimators_]
-    )
+    bootstrap_indices = np.array([
+        _generate_sample_indices(e.random_state, n_samples, n_samples) for e in est.estimators_
+    ])
     for train_idx, train_idx_prox in enumerate(proximities):
         for proximity_idx, proximity_count in train_idx_prox:
             bootstrap_count = 0
