+++ .github/workflows/publish.yml
@@ -55,6 +55,7 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
+          name: ${{ matrix.os }}-artifact
           path: ./wheelhouse/*.whl
 
   upload_pypi:
