+++ .github/workflows/publish.yml
@@ -18,69 +18,59 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
-      - uses: actions/setup-python@v5
-        with:
-          python-version: '3.12'
-
-      - name: Install cibuildwheel
-        run: python -m pip install cibuildwheel==2.16.2
-
-      - name: Build wheels for Linux
-        if: matrix.os == 'ubuntu-latest'
-        run: |
-          python -m cibuildwheel --output-dir wheelhouse
+      - name: Build wheels
+        uses: pypa/cibuildwheel@v2.16.2
         env:
-          CIBW_SKIP: "pp* *i686* *musllinux*"
+          CIBW_BUILD: cp3*-*
+          CIBW_SKIP: pp* *i686* *win32 *musllinux*
           CIBW_PROJECT_REQUIRES_PYTHON: ">=3.8"
           CIBW_BEFORE_BUILD: pip install --verbose --editable .
-
-      - name: Build wheels for Windows
-        if: matrix.os == 'windows-latest'
-        run: |
-          python -m cibuildwheel --output-dir wheelhouse
-        env:
-          CIBW_SKIP: "pp* *i686* *win32"
-          CIBW_PROJECT_REQUIRES_PYTHON: ">=3.8"
-          CIBW_BEFORE_BUILD: pip install --verbose --editable .
-
-      - name: Build wheels for macOS
-        if: matrix.os == 'macos-latest'
-        run: |
-          python -m cibuildwheel --output-dir wheelhouse
-        env:
-          CIBW_ARCHS_MACOS: "x86_64 universal2 arm64"
-          CIBW_SKIP: "pp* *i686*"
-          CIBW_PROJECT_REQUIRES_PYTHON: ">=3.8"
-          CIBW_BEFORE_BUILD: pip install --verbose --editable .
+          CIBW_ARCHS_LINUX: auto64
+          CIBW_ARCHS_MACOS: x86_64 universal2 arm64
+          CIBW_ARCHS_WINDOWS: auto64
 
       - uses: actions/upload-artifact@v4
         with:
-          name: ${{ matrix.os }}-artifact
+          name: artifact-wheels-${{ matrix.os }}
           path: ./wheelhouse/*.whl
 
+  build_sdist:
+      name: Build source distribution
+      runs-on: ubuntu-latest
+      steps:
+        - uses: actions/checkout@v4
+
+        - uses: actions/setup-python@v5
+          name: Install Python
+          with:
+            python-version: '3.12'
+
+        - run: pip install build
+
+        - name: Build sdist
+          run: python -m build --sdist
+
+        - uses: actions/upload-artifact@v4
+          with:
+            name: artifact-source
+            path: dist/*.tar.gz
+
   upload_pypi:
-    needs: build_wheels
+    needs: [build_wheels, build_sdist]
     runs-on: ubuntu-latest
 
     if: github.event_name == 'push' && startsWith(github.event.ref, 'refs/tags/v')
     steps:
-      - uses: actions/checkout@v4
-
       - uses: actions/download-artifact@v4
         with:
-          name: artifact
           path: dist
+          pattern: artifact-*
+          merge-multiple: true
 
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install build
-          pip install twine
-
-      - name: Publish
-        env:
-          TWINE_USERNAME: __token__
-          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-        run: |
-          python -m build --sdist
-          twine upload dist/*
+      - uses: pypa/gh-action-pypi-publish@v1.8.11
+        with:
+          user: __token__
[…]
