### 2024-01-02 #5976 @jbytecode
@oparisot - Thank you for accepting our invitation. Please firstly type

`@editorialbot generate my checklist`

to generate your tasklist. Thank you in advance!
