### 2024-01-02 #5976 @jbytecode
@oparisot - here is the latest compiled pdf of the submission.
