### 2024-01-09 #5976 @oparisot
Congratulations @reidjohnson for the software and the paper, it's a great job.
Thanks to the efforts made since the first reviewer's comments, I don't have much to add!
@jbytecode It deserves acceptance, without hesitation.
