### 2024-01-09 #5976 @jbytecode
@reidjohnson - The paper and the bibliography look fine. We have to meet a few requirements.

Could you please perform the tasks given in https://github.com/openjournals/joss-reviews/issues/5976#issuecomment-1883345498
