+++ .github/workflows/draft-pdf.yml
@@ -1,6 +1,6 @@
 name: paper
 
-on: [push]
+on: workflow_dispatch
 
 jobs:
   paper:
