+++ quantile_forest/version.txt
@@ -1,1 +1,1 @@
-1.2.4dev
+1.2.4
