+++ _downloads/9692ea47a09c86e0bd5b0eaab224f0dd/plot_quantile_interpolation.py
@@ -7,6 +7,7 @@ An example comparison of interpolation methods that can be applied during
 prediction when the desired quantile lies between two data points.
 
 """
+
 print(__doc__)
 
 import matplotlib.pyplot as plt
