+++ _downloads/972f82026f05bdd91ec5755a0e8a5403/plot_quantile_extrapolation_problem.py
@@ -9,6 +9,7 @@ bounds of the data in the training set, an important limitation of the
 approach.
 
 """
+
 print(__doc__)
 
 import matplotlib.pyplot as plt
