+++ _sources/sg_execution_times.rst.txt
@@ -1,0 +1,55 @@
+
+:orphan:
+
+.. _sphx_glr_sg_execution_times:
+
+
+Computation times
+=================
+**00:18.790** total execution time for 7 files **from all galleries**:
+
+.. container::
+
+  .. raw:: html
+
+    <style scoped>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
+    </style>
+    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
+    <script type="text/javascript" class="init">
+    $(document).ready( function () {
+        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
+    } );
+    </script>
+
+  .. list-table::
+   :header-rows: 1
+   :class: table table-striped sg-datatable
+
+   * - Example
+     - Time
+     - Mem (MB)
+   * - :ref:`sphx_glr_auto_examples_plot_quantile_weighted_vs_unweighted.py` (``../examples/plot_quantile_weighted_vs_unweighted.py``)
+     - 00:10.559
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_quantile_regression_intervals.py` (``../examples/plot_quantile_regression_intervals.py``)
+     - 00:05.024
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_quantile_conformalized.py` (``../examples/plot_quantile_conformalized.py``)
+     - 00:01.904
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_quantile_vs_standard_forest.py` (``../examples/plot_quantile_vs_standard_forest.py``)
+     - 00:00.540
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_quantile_extrapolation_problem.py` (``../examples/plot_quantile_extrapolation_problem.py``)
+     - 00:00.394
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_quantile_toy_example.py` (``../examples/plot_quantile_toy_example.py``)
+     - 00:00.278
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_plot_quantile_interpolation.py` (``../examples/plot_quantile_interpolation.py``)
+     - 00:00.091
+     - 0.0
