+++ _sources/auto_examples/plot_quantile_toy_example.rst.txt
@@ -26,7 +26,7 @@ An example that demonstrates the use of a quantile regression forest to
 predict a conditional median and prediction intervals. The example compares
 the predictions to a ground truth function used to generate noisy samples.
 
-.. GENERATED FROM PYTHON SOURCE LINES 11-69
+.. GENERATED FROM PYTHON SOURCE LINES 11-70
 
 
 
@@ -39,7 +39,8 @@ the predictions to a ground truth function used to generate noisy samples.
 
 
 
-.. code-block:: default
+.. code-block:: Python
+
 
     print(__doc__)
 
@@ -103,7 +104,7 @@ the predictions to a ground truth function used to generate noisy samples.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.470 seconds)
+   **Total running time of the script:** (0 minutes 0.278 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_toy_example.py:
@@ -112,16 +113,13 @@ the predictions to a ground truth function used to generate noisy samples.
 
   .. container:: sphx-glr-footer sphx-glr-footer-example
 
+    .. container:: sphx-glr-download sphx-glr-download-jupyter
 
-
+      :download:`Download Jupyter notebook: plot_quantile_toy_example.ipynb <plot_quantile_toy_example.ipynb>`
 
     .. container:: sphx-glr-download sphx-glr-download-python
 
       :download:`Download Python source code: plot_quantile_toy_example.py <plot_quantile_toy_example.py>`
-
-    .. container:: sphx-glr-download sphx-glr-download-jupyter
-
-      :download:`Download Jupyter notebook: plot_quantile_toy_example.ipynb <plot_quantile_toy_example.ipynb>`
 
 
 .. only:: html
