+++ _sources/auto_examples/plot_quantile_weighted_vs_unweighted.rst.txt
@@ -27,7 +27,7 @@ regression forest with weighted and unweighted quantiles to compute the
 predicted output values. A standard random forest regressor is included for
 comparison.
 
-.. GENERATED FROM PYTHON SOURCE LINES 12-97
+.. GENERATED FROM PYTHON SOURCE LINES 12-98
 
 
 
@@ -40,7 +40,8 @@ comparison.
 
 
 
-.. code-block:: default
+.. code-block:: Python
+
 
     print(__doc__)
 
@@ -131,7 +132,7 @@ comparison.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 18.572 seconds)
+   **Total running time of the script:** (0 minutes 10.559 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_quantile_weighted_vs_unweighted.py:
@@ -140,16 +141,13 @@ comparison.
 
   .. container:: sphx-glr-footer sphx-glr-footer-example
 
+    .. container:: sphx-glr-download sphx-glr-download-jupyter
 
-
+      :download:`Download Jupyter notebook: plot_quantile_weighted_vs_unweighted.ipynb <plot_quantile_weighted_vs_unweighted.ipynb>`
 
     .. container:: sphx-glr-download sphx-glr-download-python
 
       :download:`Download Python source code: plot_quantile_weighted_vs_unweighted.py <plot_quantile_weighted_vs_unweighted.py>`
-
-    .. container:: sphx-glr-download sphx-glr-download-jupyter
-
-      :download:`Download Jupyter notebook: plot_quantile_weighted_vs_unweighted.ipynb <plot_quantile_weighted_vs_unweighted.ipynb>`
 
 
 .. only:: html

[…]
