### 2024-01-17 #5976 @reidjohnson
@jncraton and @oparisot, thank you again for the reviews.

@jbytecode: The release version for the JOSS paper is 1.2.4. The Zenodo archive DOI is [10.5281/zenodo.10521419](https://zenodo.org/doi/10.5281/zenodo.10521419).
