### 2024-01-17 #5976 @jbytecode
@reidjohnson - Everything seems good to me. Please have a full read of the paper one more time. Please correct any issue if exists. Please ping me when you've done with them.
