+++ paper/paper.md
@@ -17,27 +17,27 @@ bibliography: paper.bib
 
 # Summary
 
-Quantile regression forests (QRF) is a non-parametric, tree-based ensemble method for estimating conditional quantiles [@meinshausen2006quantile]. It is a generalization of the random forests algorithm, a versatile ensemble learning algorithm, originally proposed in [@breiman2001random], that has proven extremely popular and useful as a general-purpose machine learning method [@biau2016random;@hengl2018random;@wager2018estimation;@athey2019generalized]. Instead of outputting the weighted mean value of training labels for regression like random forests, QRF employ the weighted empirical distribution of training labels to obtain the predictive distribution. This enables QRF to output probabilistic predictions for regression problems, which are widely useful for constructing estimates of uncertainty [@petropoulos2022forecasting].
+Quantile regression forests (QRF) is a non-parametric, tree-based ensemble method for estimating conditional quantiles [@meinshausen2006quantile]. It is a generalization of the random forests algorithm, a versatile ensemble learning algorithm, originally proposed in [@breiman2001random], that has proven extremely popular and useful as a general-purpose machine learning method [@biau2016random;@hengl2018random;@wager2018estimation;@athey2019generalized]. Instead of outputting the weighted mean value of training labels like random forests regressors, QRF employs the weighted empirical distribution of training labels to obtain the predictive distribution. This feature enables QRF to output probabilistic predictions for regression problems, which are widely useful for constructing estimates of uncertainty [@petropoulos2022forecasting].
 
-`quantile-forest` provides a fast, feature-rich implementation of the QRF algorithm. The estimators provided in this package are optimized using Cython [@behnel2010cython] for training and inference speed, and can estimate arbitrary quantiles at prediction time without retraining. They provide methods for out-of-bag estimation, calculating quantile ranks, and computing proximity counts. The provided QRF estimators are also compatible with and can serve as drop-in replacements for the widely used forest regressors available in scikit-learn [@kramer2016scikit]. The package was designed to be used by a broad array of researchers and in production business settings. It has already been cited in scholarly work [@althoff2023conform;@saporta2023statistical;@Prinzhorn2023] and used in a production setting at the real estate technology company Zillow. The combination of speed, design, and functionality in `quantile-forest` enables exciting scientific explorations in academia and industry alike.
+`quantile-forest` provides a fast, feature-rich QRF implementation. The estimators provided in this package are optimized using Cython [@behnel2010cython] for training and inference speed, and can estimate arbitrary quantiles at prediction time without retraining. They provide methods for out-of-bag estimation, calculating quantile ranks, and computing proximity counts. The provided QRF estimators are also compatible with and can serve as drop-in replacements for the widely used forest regressors available in scikit-learn [@kramer2016scikit]. The package is designed to be used by a broad array of researchers and in production business settings. It has already been cited in scholarly work [@althoff2023conform;@saporta2023statistical;@Prinzhorn2023] and used in a production setting at the real estate technology company Zillow. The combination of speed, design, and functionality in `quantile-forest` enables exciting scientific explorations in academia and industry alike.
 
 # Statement of Need
 
 Quantile regression is useful for understanding relationships between variables outside of the mean of the data, which can be particularly useful for understanding outcomes that are non-normally distributed or that have nonlinear relationships with predictor variables. It can be used to understand an outcome at its various quantiles and to compare groups or levels of an exposure on those quantiles [@koenker2005quantile].
 
[…]
