### 2024-01-18 #5976 @reidjohnson
@jbytecode: I made some minor corrections/updates to the paper. No claims, statements, or references were changed (full diff [here](https://github.com/zillow/quantile-forest/commit/2221f4e243c92890a2c0769daca0b689ab0f7b99)). With that, the paper looks good to me.
